LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_navigation_download_block_downloads.sub.tentative-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing


FAIL Navigation resulted download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
PASS Navigation resulted download in sandbox from <object> is blocked.
